Finished move.h
Added <fcntl.h> directive for renameat.
Finalized the logic for creating folders in the user's home.
In that logic, there should be options for picking where to save and
what to name the bash script created by this program.
--NOTE--
If anyone is reading this, yeah it's not quite "native" I guess,
but that's how I named it for simplicity's sake while coding.